requirements.txt
streamlit
openai
fpdf2
pypdf
Pillow
requests
beautifulsoup4
lxml